Remove outdated enrollment and graduation data scripts; add new country statistics CSV and gender plotter script for enhanced analysis
