--- synthetic datasets/degree_enrollment.csv
@@ -1,16 +1,0 @@
-year,CS-MAJO,CS-SPEC,CS-MANA,IT-MAJO,IT-SPEC,total
-2007,3,0,26,29,0,58
-2009,2,0,5,51,0,58
-2010,22,0,13,61,0,96
-2011,2,21,16,72,0,111
-2012,4,43,15,54,0,116
-2013,2,31,8,81,0,122
-2014,13,17,4,65,0,99
-2015,8,48,9,51,0,116
-2016,30,162,30,127,94,443
-2017,33,201,42,89,91,456
-2018,72,196,41,74,101,484
-2019,130,173,67,111,93,574
-2020,170,167,82,125,76,620
-2021,207,174,103,131,66,681
-2022,264,187,99,132,61,743
